refactored folder structure, refactored nav links, moved routes from index.js to their own file, fixed issue with solar system planets not following orbit lines
