+++ client/src/App.css
@@ -1,0 +1,35 @@
+.App {
+	display: flex;
+	text-align: center;
+	overflow: hidden;
+	position: absolute;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 100vw;
+	height: 100vh;
+	background: black;
+	/* stars */
+	background-image: radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1.5%),
+		radial-gradient(circle at center, white 0%, transparent 2%), radial-gradient(circle at center, white 0%, transparent 1.5%),
+		radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1.5%),
+		radial-gradient(circle at center, white 0%, transparent 2%), radial-gradient(circle at center, white 0%, transparent 1.5%),
+		radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1%);
+	background-size: 100px 100px;
+	background-position: -50px 0px, 10px 20px, -26px -45px, 12px -12px, -12px 40px, 30px 48px, 30px -20px, -20px -25px, -20px 10px, 55px 20px;
+}
+.App::before {
+	content: "";
+	position: absolute;
+	width: 100%;
+	height: 100%;
+	background-image: linear-gradient(5deg, rgba(0, 0, 0, 0.4) 30%, rgba(255, 0, 0, 0.1), rgba(0, 0, 0, 0.4) 60%),
+		linear-gradient(-15deg, rgba(0, 0, 0, 0.4) 30%, rgba(0, 255, 0, 0.1), rgba(0, 0, 0, 0.4) 70%),
+		linear-gradient(rgba(0, 0, 0, 0.2), rgba(0, 0, 255, 0.15), rgba(0, 0, 0, 0.2));
+}
+
+.overlay {
+	position: absolute;
+	height: 100vh;
+	width: 100vw;
+}
