+++ client/src/components/Details/details.css
@@ -1,6 +1,3 @@
 .details {
-	left: 60vw;
-	position: absolute;
-	z-index: 20;
 	color: white;
 }
