+++ client/src/components/navbar/style.css
@@ -1,4 +1,6 @@
 .nav {
+	position: absolute;
+	top: 0;
 	display: flex;
 	justify-content: space-between;
 	color: white;
