+++ client/src/pages/Planet Page/style.css
@@ -3,5 +3,5 @@
 	display: flex;
 }
 .planet-page section {
-	border: 4px solid white;
+	width: 50%;
 }
