--- client/src/routes/root.css
@@ -1,71 +1,0 @@
-.App {
-	display: flex;
-	width: 100vw;
-	min-height: 100%;
-	text-align: center;
-	overflow: hidden;
-}
-
-.App-logo {
-	height: 40vmin;
-	pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-	.App-logo {
-		animation: App-logo-spin infinite 20s linear;
-	}
-}
-
-.App-header {
-	background-color: #282c34;
-	min-height: 100vh;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	font-size: calc(10px + 2vmin);
-	color: white;
-}
-
-.App-link {
-	color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-	from {
-		transform: rotate(0deg);
-	}
-	to {
-		transform: rotate(360deg);
-	}
-}
-
-.planet-display {
-	position: absolute;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: 100vw;
-	height: 100vh;
-	background: black;
-	/* stars */
-	background-image: radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1.5%), radial-gradient(circle at center, white 0%, transparent 2%), radial-gradient(circle at center, white 0%, transparent 1.5%), radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1.5%), radial-gradient(circle at center, white 0%, transparent 2%), radial-gradient(circle at center, white 0%, transparent 1.5%), radial-gradient(circle at center, white 0%, transparent 1%), radial-gradient(circle at center, white 0%, transparent 1%);
-	background-size: 100px 100px;
-	background-position: -50px 0px, 10px 20px, -26px -45px, 12px -12px, -12px 40px, 30px 48px, 30px -20px, -20px -25px, -20px 10px, 55px 20px;
-}
-
-/* aurora */
-.planet-display::before {
-	content: '';
-	position: absolute;
-	width: 100%;
-	height: 100%;
-	background-image: linear-gradient(5deg, rgba(0, 0, 0, 0.4) 30%, rgba(255, 0, 0, 0.1), rgba(0, 0, 0, 0.4) 60%), linear-gradient(-15deg, rgba(0, 0, 0, 0.4) 30%, rgba(0, 255, 0, 0.1), rgba(0, 0, 0, 0.4) 70%), linear-gradient(rgba(0, 0, 0, 0.2), rgba(0, 0, 255, 0.15), rgba(0, 0, 0, 0.2));
-}
-
-.overlay {
-	position: absolute;
-	height: 100vh;
-	width: 100vw;
-}
